make quitting on a non-empty reward account forbidden
The Cardano ledger forbids to de-register a stake key, if there are
still rewards associated with it. We could have implemented some more
convoluted logic where the wallet would automatically collect rewards as
part of a de-registration, but we are a bit lacking time here so we'll
simply make this a manual operation that users have to undergo. Yet, the
backend will still return a meaningful error message explaining the
situation.