consume withdrawal gradually, instead of spliting them proportionally
It is a bit more work to get this right, but in the end, it solves the same problem (being able to use the withdrawal on several outputs) with more elegance and efficiency.
Now the withdrawal is used gradually, and only what's needed at every step so in that sense, it is optimal.