increase the delta between small and big state when comparing them in test
It seems that we are getting very small rewards in CI (the root cause is unsure, but it could be that pools are punished for missing slots). A difference may not be noticeable if the stake difference is insufficient for this test. So increasing the gap even further to increase the chance of finding a difference