Revert "Rename sqlite `slot` fields to `slot_no` forcing migration"
This reverts commit 9eb843b6075d92f8a149479f9d8f9b4c5671d98e.

That commit broke a unit test failure for a manual migration. I am not
sure if the test could have been removed without consequences.

But weirdly, migration seems to work smoothly without it:

$ cardano-wallet-shelley-2020-07-06 serve --node-socket socket --testnet shelley_testnet-genesis.json --database wallet-db
$ # Create a wallet
$ cardano-wallet-shelley transaction list <WALLET ID>

$ cardano-wallet-shelley serve --node-socket socket --testnet shelley_testnet-genesis.json --database wallet-db
$ cardano-wallet-shelley transaction list <WALLET ID>

$ # The correct slots are still shown.

Looking in the DBs, it seems like wallet does detect that the schema has
changed, and re-syncs from genesis.