Ensure that no test stake pools retire prematurely.
For now, we actually want all test stake pools to remain alive for the
full duration of an integration test suite run, hence we configure our
test pools to have retirement dates that are well into the future.

In future, we will add more tests to simulate what happens after a stake
pool has retired.

In response to review feedback:

https://github.com/input-output-hk/cardano-wallet/pull/1931#discussion_r457417183