Report all pool metric values as zero when real pool LSQ data is absent.
When a pool is registered (with a registration certificate) but not yet
active (and therefore not yet causing pool metrics data to be available
over local state query), we would nevertheless still prefer to include
that pool in the list returned by the `ListStakePools` API endpoint.

This change causes such pools to be listed, and reports all pool metric
values as **zero** when pool LSQ data is absent.