Don't fetch nOpt another time
nOpt is returned as part of the `stakeDistribution` function. It was
originally added for the purpose of calculating saturation.

Recently, along with logic for handling pools not-yet in the stake
distribution, this was added:
```
pp <- liftIO $ getProtocolParameters nl
```
also for the sake of retrieving nOpt (aka desiredNumberOfStakePools).

It is both slightly confusing to retrieve it from two different places,
and they are no longer guaranteed to be from the same point.