Don't store GenesisParameters in checkpoints
Most GenesisParameters are unneeded in the checkpoints, and storing them
there may give the impression that they may be changed by protocol
updates and forks. They won't. (There is a separate ProtocolParameters
table, however)

This commit keeps the genesis hash in the checkpoints, such that we
still can validate the integrity of the db.

For pruning checkpoints, the db needs the security parameter k. Since we
no longer can read it from the DB, we pass it in when constructing the
db.