Batch getAccountBalance from multiple workers together
getAccountBalance is called by each wallet worker, I think on each
restoration step (which is often).

For some reason, sending multiple getAccountBalance queries at the same
time slows it down from e.g. 0.002s to 40s.

The LSQ query even supports querying multiple accounts in a single
query, so let's use it!

This commit makes getAccountBalance merely lookup in a Map in a TVar,
which is updated with a single query on every tip-change.

Somewhat crude implementation so far.

I'm yet not familiar with the code that is calling getAccoutBalance. I
suspect there are layers of indirection that can be removed.