Track registrations separately from cache
Before this commit keys would start in `toBeObserved`. They would then
get passed to the `fetch` function. If the supplied `fetch` function
chose to not include the same keys in the resulting map of values, they
would silently dissappear.

This way, `newObserver` takes full responsibility of maintaining the set
of keys which should be observed, by replacing the TVars (cache,
toBeObserved) with (cache, observedKeys)