Use `RecordWildCards` in `newDBLayer` functions.
Justification:

Some `DBLayer` functions are defined in terms of other `DBLayer`
functions, such as `readPoolLifeCycleStatus`, which is defined in terms
of `readPoolRegistration` and `readPoolRetirement`.

Since mutual recursion between functions defined as fields of a record
is somewhat awkward, we currently define auxilliary functions for each
of the functions that we need to reuse, and then refer to these
auxilliary functions when building the record. However, this creates a
layer of indirection and duplication.

By using `RecordWildCards`, we can instead build a `DBLayer` object in
terms of ordinary functions defined in a `where` clause. Since ordinary
functions can refer to one another using mutual recursion, we no longer
need auxilliary functons, saving a layer of indirection.