Remove nested call to `atomically`.
Nesting a call to `atomically` will result in the nested call blocking
indefinitely.

This fixes:

https://github.com/input-output-hk/cardano-wallet/pull/2057#issuecomment-680825458
https://github.com/input-output-hk/cardano-wallet/pull/2057#issuecomment-680833124

This change also removes bracketing for the `MsgStakePoolGarbageCollection`
constructor of `StakePoolLog`, as there are no longer any child log entries
belonging to the pools engine.

(Other log entries pertaining to garbage collection are emitted by the pool
DB component, as a result of calling the `removeRetiredPools` DB function.)