Fix the model implementation of `listRetiredPools`.
The `prop_poolRetirement` property expects that `listRetiredPools` will
include all pools with non-cancelled retirement certificates, even when
those pools never had registration certificates to begin with.

Strictly speaking, this expectation is more than we need, as in reality
we'll probably never see a case where a pool publishes a retirement
certificate without there being a prior registration certificate. (This
should be explicitly forbidden by the ledger layer.)

However, it turns out to be more convenient to express the property this
way, as:

1. It makes the property much simpler to express. We don't need to worry
   about registration certificates in the setup phase.
2. It makes the SQLite query for implementing `listRetiredPools` much
   easier to write.

Therefore, in this case, we simply update the model to match the
behaviour of our SQLite implementation.