Fix the model implementation of `mUnfetchedPoolMetadataRefs`.
The property expects that recent fetch attempts are excluded from the
list returned by `unfetchedPoolMetadataRefs`. We adjust the model to
satisfy the property.

Note that the property coverage here isn't necessarily as complete as it
could be. We'd ideally want to test this pool DB function in conjunction
with other pool DB functions as part of a quickcheck-state-machine style
test.