Use `List.sort` instead of `Set.fromList`.
The `listRegisteredPools` function doesn't guarantee any particular
order, but it should guarantee to not return any duplicate elements.

Use `List.sort` here instead of `Set.fromList` to avoid masking this
potential error.

(Alternatively, we could just adjust the type of `listRegisteredPools`
to return a set instead of a list, but that is a more invasive change.)