Limit pool garbage collection to once per epoch.
Before this change, it was possible for garbage collection to happen
more than once per epoch, as any rollback would recreate the mutable
reference to the latest garbage collection epoch, causing the latest
garbage collection epoch value to be reset to zero.

This change adjusts the reference creation mechanism so that it only
happens once, on startup.