Test that `removePools` does not remove metadata.
Strictly speaking, when removing pools, we should also be able to remove any
pool metadata that becomes unreachable as a result.

However, this is tricky to get right, since it's possible for more than one
pool to share the same metadata hash.

For example, if two pools p and q issue registration certificates that share
the same metadata hash h, and then we garbage collect pool p, we have to be
careful to not remove the metadata corresponding to h in order to preserve
the metadata for pool q.

One possible solution would be to remove metadata entries only when we can
show that they are no longer reachable: when metadata entries are no longer
referenced by any pool.

However, for the moment, we adopt the principle of never removing metadata.
We can revise this decision in future if the metadata table grows too large.