Modify several `Buildable` instances to include tokens.
This change modifies the `Buildable` instances of `Wallet`, `UTxO`, and
`TxOut` so that they include the contents of their inner token bundles.

This is important for debugging test failures, as we sometimes need to be
able to inspect and compare the token bundles of two transaction outputs.

The output for an individual `TxOut` value is now of the following form:

```
- input: 2nd 66646130
  output:
    address: 2a0f753f...6e6a7064
    coin: 21
    tokens:
      - policy: aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa
        tokens:
          - token: TokenC
            quantity: -704
          - token: TokenD
            quantity: 256
      - policy: bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
        tokens:
          - token: TokenA
            quantity: -16
          - token: TokenB
            quantity: 72
          - token: TokenC
            quantity: 392
          - token: TokenD
            quantity: 360
```