Don't cache core/shelley unit test results for Hydra bors jobsets
The unit tests are getting a bit flaky, like the integration
tests. The small amount of time saved from caching unit test runs is
not worth it. It's annoying when resubmitted bors jobs fail due to
cached test results.

This change disables caching of unit test results for the
cardano-wallet-core and cardano-wallet packages, if the jobset is part
of a bors run.