Switch to using UnliftIO.Exception bracket
It turns out that the Control.Exception.bracket function isn't very
safe to use with async code. The reason is the bracket cleanup handler
may be interrupted by async exceptions, and if that happens, then
cleanup action may not be done.

I am wondering whether this may be the cause of some test
flakiness. Also I wonder whether cleanup handlers are the cause of
deadlocking tests.

This change replaces Control.Exception with UnliftIO.Exception - the
"async safe" variant.

The other benefit of MonadUnliftIO is that it provides "unlift" for
functions of the form withFoo :: (foo -> IO a) -> IO a -- i.e. it lets
you replace IO with other kinds of m.

Docs here: https://github.com/fpco/unliftio/tree/master/unliftio#readme

So it may fix flaky tests and/or help to avoid future concurrency
bugs.