Re-add Qry wrapper for safe composition
We had missed that HF.Qry by design only works within a single era.

When integrating with Shelley we wrapped HF.Qry in our own Qry type.
This hid the problem.

When integrating with Allegra, we replaced the Qry wrapper with the
simpler `type Qry = ReaderT StartTime HF.Qry`. This caused ADP-626 -
transactions couldn't be listed if the filter time ranges spanned
multiple eras.

This commit re-adds the Qry wrapper, and makes the failing regression
test (previous commit) pass.

Is this the optimal solution? I don't know. The alternative would be
forcing slotRangeToTimeRange to be a function rather than a @Qry@.