.gitattributes: Minimize golden test JSON files in GitHub PR
This causes golden JSON files used for tests to be hidden by default
in the GitHub PR review page. They are hidden in the same way that
"large diffs" are hidden by default.

The reviewer can then concentrate on the substance of the PR, and then
inspect JSON files by clicking on the "Load diff" link.