Remove unsafe constructors for `TokenName` and `TokenPolicyId`.
This change removes the smart constructors for the `TokenName` and
`TokenPolicyId` types.

In the case of `TokenPolicyId`, this constructor was actually unsafe.

The only place these functions were actually used was in the test code,
which has now been updated to use `fromText`.

In future, we can provide a function that produces a `TokenPolicyId`
value by hashing the content of a token policy.

Fixed in response to review feedback:

https://github.com/input-output-hk/cardano-wallet/pull/2384#discussion_r540810284