Change the inner type of `TokenQuantity` to `Natural`.
At present, the only use case for these quantities is to be included in
token bundles held within transaction outputs, and these must never be
negative.

When we build support for minting and burning of tokens, we may wish to
parameterize this type and allow it to be instantiated with 'Integer'.

In response to review feedback:

https://github.com/input-output-hk/cardano-wallet/pull/2384#discussion_r544243932