Set CARDANO_NODE_SOCKET_PATH only in child processes
Using the global environment variable CARDANO_NODE_SOCKET_PATH caused
problems when setting up multiple test clusters in one process.

So now the socket path is passed only to the cardano-cli commands
which need it.

The bulk of this change is using the CardanoNodeConn newtype
everywhere, so that FilePath arguments don't get mixed up.

Then, the `cli` function calls needed to be modified to set
CARDANO_NODE_SOCKET_PATH in the child process environment, if
needed. Since I was changing them all anyway, I thought we may as well
use typed-process, which is better and safer than process.