Make it possible to limit the size of a selection.
It's important to be able to protect the wallet from computations that
are excessively costly.

This change introduces the `SelectionLimit` type, which makes it possible to
limit the scope of a coin selection computation by specifying an upper limit
on the number of inputs that can be selected.

With this change, the coin selection algorithm will terminate when the limit
is reached, and not make any further selections.

If, during a selection run, the limit is reached, there are two main cases:

  - If the current selection already satisfies the minimum required
    balance, `performSelection` will return the current selection.

  - If the current selection does not satisfy the minimum required
    balance, `performSelection` will return a `SelectionInsufficient`
    error.