Clarify the distinction between user-specified and non-user-specified assets.
The user-specified criteria for a coin selection include a set of outputs.
These outputs, when considered collectively, correspond to a well-defined
set of assets, which we refer to as "user-specified assets".

When performing random selection of UTxO entries to cover a given asset, we
may end up selecting entries that contain additional assets: assets that
did not appear in the original outputs. Since these assets were not
specified by the user, we refer to them as "non-user-specified assets".