Simplify LSQ by removing getTip
Simplify the local state query client by having it acquire the /latest/
tip (`Nothing`), instead of asking for a specific tip from the tip sync
client.

To know the era, we now need to send a query. But this arguably just
makes things neater. It also resolves a problem from the previous
commits, where LSQ would block on wallet startup until the node changed
tip (could be on the order of 10s), since the tip sync client only had
received a MsgRollback without information of the current era.