Fork `computeMinimumAdaQuantity` into public and internal variants.
The public function accepts a `TokenMap`, which is what the coin
selection algorithm expects to be able to supply.

The internal function, which is a simple wrapper around the ledger
specification's `scaledMinDeposit` function, still accepts a `TokenBundle`.

This is because we still need to test that this function is agnostic
w.r.t. ada quantities.