Finish moving core-integration
We need to share integration test code between
cardano-wallet:integration and cardano-wallet:bench:latency.

We want the test code to depend on cardano-wallet.

We don't want to pollute the cardano-wallet library with the test code.

So this commit puts it in a new internal library, integration-framework.

This causes problems with weeder. If going down this root, we probably
need to upgrade to weeder 2 (https://github.com/ocharles/weeder).

TODO: Maybe we could move Shelley.Cluster to the integration-framework
package.