Remove redundant eventually
The fork is at genesis. If the node has synced anything (which we should
be able to expect), then we shouldn't need the eventually.

A recent change in the NetworkLayer also causes the wallet to block
until it can retrieve the node's PParams, further decreasing the risk of
something unexpected being returned here.