Add unit tests for the maximum token quantity.
This change adds unit tests for `performSelection` to demonstrate what
should happen when inputs are selected whose total quantity of a given
asset is close to the maximum token quantity.

It should not be possible to produce a `SelectionResult` where one or
more change outputs contains a token quantity greater than the maximum
(maxBound :: Word64).

These unit tests serve as a simple sanity check: the primary testing of
the change map partitioning behaviour is handled by the following
property tests:

 - prop_equipartitionCoin_*
 - prop_equipartitionTokenBundle_*
 - prop_equipartitionTokenMap_*