When assessing the size of a change map, use the maximum ada quantity.
When assessing the size of a change map to determine if it is
excessively large, we don't yet know how large the associated
ada quantity will be, since ada quantities are assigned at a
later stage (in 'assignCoinsToChangeMaps').

Therefore, we err on the side of caution, and assess the size
of a change map combined with the maximum possible ada quantity.

This means that when presented with a very large change map, we
have a small chance of splitting the map even if that map would
be within the limit when combined with its final ada quantity.

However, oversplitting a change map is preferable to creating
a bundle that is marginally over the limit, which would cause
the resultant transaction to be rejected.

In response to review feedback and discussion:

https://github.com/input-output-hk/cardano-wallet/pull/2552#discussion_r589335232