Use `FollowLog innerTr` instead of the other way around
Turn it around: have chain follower be on the form `FollowLog
WalletWorkerLog` rather than having `WalletWorkerLog` include `FollowLog`.

The goal of this is to be able to write more useful follow-logs in a way
that benefits both wallet and pool followers. After this commit, we
should be able to have `follow` discard WalletWorkerLogs under normal
circumstanses, but flush a small buffer if something goes wrong. Or
things in that direction.