Move logic into follow from call-sites
Goal: have the follow function itself be responsible to handle rollbacks
and exceptions. If follow is only called /once/ designing good logs
should be easier. The follow function can then provide more information
in logs, as it knows what it will do on e.g. exceptions.

Move Tip and SyncProgress tracing to follow

We will re-add the syncProgress tracing shortly a as part of the
log aggregation