Add syncProgress function to NetworkLayer
By implementing it in the NetworkLayer we are able to have it return
NotResponding when we initially don't have a TimeInterpreter on startup.

This is useful to be able to log syncProgress even if the node is not
responding.

NOTE: For simplicity, this commit removes the distinction between the
wallet and node sync tolerances in the restore bench.

The tolerance used to be 300s for waitForNodeProgress, but now we use
the one provided by the NetworkLayer, which we set to 3600s.