Strengthen the `totalFee` check within `Planning.prop_createPlan`.
Before this commit, to check the correctness of `totalFee`, we simply
summed the `fee` field values of the individual selections, and checked
this value matched the total.

Now, we manually compute the expected total fee by finding the
difference between:

    1. the total amount of ada provided by the inputs and reward balance;
    2. the total amount of ada spent by the outputs.

This is a stronger test, as we avoid relying on the `Selection` module's
guarantee that the fee of each selection is correct.