[Refactoring] Introduce `TxSkeleton` type.
This record type includes the minimal set of fields that are absolutely
necessary for the `estimateTxSize` function to perform its calculations,
and nothing else.

As such, the fields of `TxSkeleton` contain a subset of the data included
in the union of `SelectionSkeleton` and `TransactionCtx`.

Using this type means that future tests for `estimateTxSize` will only
need to supply values that will actually be used by the function,
simplifying the effort to write tests.