Report fee excess in `prop_{create,extend}`.
For any selection created by `create` or `extend`, we require that the
fee excess cannot be minimized further. So we check the following
condition:

>>> feeExcess == feeExcessExpected

If this condition is ever violated in a test, it's important to know
what the actual value of `feeExcess` was.

(The value of `feeExcessExpected` is calculated by attempting to
minimize the fee excess a second time: this should have no effect, since
`minimizeFee` is designed to be idempotent.)