Remove inappropriate checks from `prop_create` and `prop_extend`.
These checks are not only unnecessary, they are incorrect.

The original motivation for including them was to check that `create`
and `extend` actually do minimize the fee (rather than leaving it
unbalanced).

But this can be adequately checked by verifying that the fee excess
cannot be minimized further, i.e., that `minimizeFee` has no effect,
which we already do.

Furthermore, these checks are incorrect, as `balance` expects to be
provided with outputs with ada quantities that are minimal.