Create "migrations/plan" endpoint for creating migration plans.
The migration planning endpoint will need to accept some data: the set
of addresses to which the wallet balance should be migrated. Without a
set of addresses, it's impossible to construct the coin selections
required for a migration plan, as every coin selection output must have
an address: these can only be provided by the caller.

However, the "migrateByronWallet" and "migrateShelleyWallet" operations
at the "migrations" endpoint already accept POST data.

We need a way to discriminate between the following operations:

1. Creating a plan.
2. Creating a plan and executing it immediately.

Assuming both types of operation require data to be POSTed, this means
we need different endpoint paths.