Rename `getMigrationInfo` to `createMigrationPlan`.
This family of API operations will eventually change from HTTP GET to
HTTP POST, as each operation requires a list of addresses to be sent
within the request body.

Therefore, it no longer makes sense to use the `get` prefix.

Additionally, since the result returned is an entire migration plan and
not just a few pieces of information, we might as well change the "Info"
suffix to a "Plan" suffix. (This naming is consistent with the
already-established concept of creating a migration plan.)