Fix termination condition in `makeChangeRepeatedly`.
This initial termination condition was incorrect:

```
    Right change | length change == length outputsToCover ->
        -- We've succeeded in making the optimal number of change outputs.
        -- Terminate here.
        pure $ Right $ mkSelectionResult change
```

It was based on the faulty assumption that `makeChange` always returns a
list of change outputs whose length is less than or equal to the number
of user-specified outputs.

However, this assumption is *NOT* valid.

In cases where there are many unique assets in the selected inputs, it's
possible for `makeChange` to return a list whose length is greater than
the number of user-specified outputs. This can happen when one or more
change outputs exceeds either of the following limits:

    1. A change output has a quantity of a particular asset that exceeds
       maxBound :: Word64.

    2. A change output has a serialized length that is greater than the
       upper limit defined by the protocol (4000 bytes).

If `makeChange` encounters a bundle that exceeds either of the above
limits, it will split up the oversized bundle into smaller bundles that
don't exceed the limits.

This means that `makeChange` may return more change outputs than the
number of user-specified outputs. Hence, we need to adjust the
termination condition to allow for this possibility.

The original justification for generating one change output per
user-specified output was to avoid shrinking the size of the UTxO
unnecessarily.

By default `makeChange` will already do this, unless there is not enough
ada available in the existing set of selected inputs to pay for all the
change, in which case change outputs will be coalesced.