Make test cluster epoch length twice as long
In some stake pool tests we use `waitForNextEpoch` to ensure some
requests and assertions occur within the same epoch.

Recent reward-related CI failures suggest that they experience an epoch
changes regardless.

To try to alleviate the problem, this commits doubles the epoch length.

The downside is that tests could take longer. With -j 6 locally I see
no meaningful regression. With -j 3 in CI there could maybe be, but if
it provides better stability it would be worth it regardless.