Mark LoggingSpec pending on macOS
This might help reduce the unit test timeouts on macOS hydra builds.

Investigating #2472 Jonathan found the following tests to fail with limited ulimit -v:
    36% Match String: "Checkpoint"
    44% Match String: "correct time measures"
   100% Match String: "different request ids"
    20% Match String: "Private Key"
    32% Match String: "Tx History"
    28% Match String: "Wallet Metadata"

where entries 2 and 3 are from LogginsSpec.

The timeouts in CI were in these following locations:

  12 times (28%) 000000 // from "Not Allowed Methods"
  9 times (21%) Checkpoint
  8 times (19%) rollback
  5 times (12%) 8601 // from e.g. "ISO 8601 extended format without timezones"
  4 times (9%) JSON
  4 times (9%) MVar
  4 times (9%) readStakeDistribution // only in the past, so now fixed
  1 times (2%) Coverage

which either suggest there are an 10+ lines of output not shown in the
output, or the LoggingSpec isn't to blame.

So it would be very interesting to see whether this commit reduces
timeouts or not.