Remove readWalletMeta from restoreBlocks
`readWalletMeta` has a hidden dependency on un up-to-date TimeInterpreter.
Under heavy load, on the local cluster with short horizons, it seems to
occasionally not be up to date.

The removed log message had Debug severity, so keeping it shouldn't be
that important.

Outstanding problems:
- Technically the pool follower also depends on an up-to-date
TimeInterpreter. Hopefully problems here are less likely than for the
pool worker.
- `readWalletMeta` should either never fail with PastHorizon errors, or
make the possibility clear to followers.