Quieten `Show` instances for API types.
Deriving a `Show` instance for newtype `T` via `Quiet T` allows the
deconstructor function for `T` to be omitted in the output of `show`.

With stock deriving, we often obtain very verbose output, similar to the
following:

```hs
derivationPath = ApiT
    { getApiT = DerivationIndex
        { getDerivationIndex = 2147485500 }
    } :|
    [ ApiT
        { getApiT = DerivationIndex
            { getDerivationIndex = 2147485463 }
        }
    , ApiT
        { getApiT = DerivationIndex
            { getDerivationIndex = 2147483648 }
        }
    , ApiT
        { getApiT = DerivationIndex
            { getDerivationIndex = 0 }
        }
    , ApiT
        { getApiT = DerivationIndex
            { getDerivationIndex = 4 }
        }
    ]
```

This occasionally causes integration test error output to be extremely verbose.

But when deriving via `Quiet`, the above output becomes:

```hs
derivationPath =
    ApiT (DerivationIndex 2147485500) :|
    [ ApiT (DerivationIndex 2147485463)
    , ApiT (DerivationIndex 2147483648)
    , ApiT (DerivationIndex 0)
    , ApiT (DerivationIndex 4)
    ]
```