Fix `estimateMaxNumberOfInputs` test.
In this test, the expected numbers of inputs are highly sensitive to the
size distribution of token bundles within generated transaction outputs.

This commit imposes a limit on the sizes of generated token bundles, so
that the expected numbers of inputs are as close as possible to the
previous set of values.