Replace deleteCascadeWhere with deleteCascade
Replace deleteCascadeWhere with deleteCascade

deleteCascadeWhere is deprecated with the suggestion:
> You can now set cascade behavior directly on entities in the quasiquoter.
https://hackage.haskell.org/package/persistent-2.13.0.2/docs/Database-Persist-Class-DeleteCascade.html

We do have OnDeleteCascade in our TH files. So it sounds like this
should still cascade correctly.

I also tested manually on alonzo testnet, with a custom wallet build for
access to W.deleteWallet (normally the entire db is deleted, but this
function is not called)

1. Restore wallet with funds
2. Inspect sqlite DB manually
3. W.deleteWallet
4. Inspect sqlite DB manually

The results were:
In step 2 there were several checkpoints. In step 4 they were gone.

W.deleteWallet does not directly delete anything from the checkpoints
table.

So it seems to work.