Add more ranking information to StakePoolSummary
Introduces and populates new fields in the GET /stake-pools endpoint.

- `metrics` > `desirability_score` provides an additional score for ranking pools with zero `non_myopic_member_rewards`
- `metrics` > `owner_stake` records the stake delegated by a stake pool owner in order to meet his/her pledge.
- The convenience flag `owner_stake_lower_than_pledge` highlights when a stake pool does not meet his/her pledge.

WARNING: The implementation in this commit may be too slow for mainnet.

Note: The `performanceEstimate` is not a `Percentage`, but a `Rational`. Pools may produce more blocks than expected due to the random nature of the block production lottery. In fact, if there is a pool that had bad luck in the lottery, then there must necessarily be at least one with pool with better than average luck.