Add libstdc++-6.dll and other DLLs to windows packages
This is rather messy. gcc puts it's libraries into a architecture
specific subdirectory which makes this rather fragile.  On top of
that the library we depend on brings in more dependencies.

We have a slightly similar situation with remote-iserv, which we
can find in the remote-iserv.postInstall of the haskell.nix
overlays/windows.nix.  The solution there is to just take the
glob all approach and find all .dlls and copy them, not the
prettiest but a workable for a tool like remote-iserv.  In this
case we might want to do something smarter akin to the following
pseudo code:

  candidatePackage = [ pkgs.libsodium-vrf pkgs.buildPackages.gcc.cc ... ]
  knownDlls = concat [ find pkg -name "*.dll" | pkg <- candidatePackage ]
  projectDlls = <paths to all already pacakged dlls and exes>
  for dll in projectDlls; do
    for dep in $(strings dll | grep ".dll"); do
      match = grep "$dep\$" knownDlls
      if match; then
        ln -s $match $out/bin
      fi
    done
  done