Test Mary works without TestEnableDevelopmentNetworkProtocols
Instead of always setting TestEnableDevelopmentNetworkProtocols in the
integration tests, with this commit we only set it if the era is Alonzo.

Because we run the integration tests on Mary in buildkite, this should
mean that the wallet now again works with mainnet nodes, without
TestEnableDevelopmentNetworkProtocols.

Haven't actually tested to connect against mainnet. But I did sanity
check that we got

    "eras": {
        "shelley": {
            "epoch_start_time": "2021-07-31T19:10:38Z",
            "epoch_number": 0
        },
        "mary": {
            "epoch_start_time": "2021-07-31T19:10:38Z",
            "epoch_number": 0
        },
        "byron": {
            "epoch_start_time": "2021-07-31T19:10:38Z",
            "epoch_number": 0
        },
        "allegra": {
            "epoch_start_time": "2021-07-31T19:10:38Z",
            "epoch_number": 0
        },
        "alonzo": null

from GET /network/parameters

when running `LOCAL_CLUSTER_ERA=mary NO_POOLS=1 NO_CLEANUP=1 local-cluster`

and that TestEnableDevelopmentNetworkProtocols was absent from the BFT
node's config.