Simplify type of function `asCollateral`.
This commit simplifies the type of `asCollateral` to just take a `TxOut`.

The value of the `TxIn` parameter was always ignored, so it can be
safely removed from the input of the function.

This enables us to make some simplifications to the test suite.

It won't affect the usability of the `asCollateral` function, since we can
now call it like this:

```hs
utxoCollateral :: UTxO -> Map TxIn Coin
utxoCollateral (UTxO u) = Map.mapMaybe asCollateral u
```