Add missing Alonzo-case in decodeSignedTx
1. Add missing Alonzo-case in decodeSignedTx, breaking the /v2/proxy/transactions
   endpoint.
2. Replace the wildcard `_` with a case for Byron, such that we don't
   repeat the mistake for future eras.

   Although, the decodeSignedTx will be simplified before then anyway.

We used to have integration tests for external tx submission, but they
were removed in #2612 to drop the cardano-transactions dependency.

This fix is not yet tested. I believe, with the tx rework, all txs will
be submitted through the external endpoint anyway. So then we will have
good coverage. For now perhaps a quick check is enough?