Replace broken shortfall calculation in `makeChange`.
In the case where `makeChange` returns an `UnableToConstructChange`
error, it attempts to record the exact shortfall in the error record.
However, this shortfall calculation was incorrect in the case where
change bundles are oversized and must be split.

Instead of attempting to calculate the shortfall within `makeChange`
itself, we instead ask `assignCoinsToChangeMaps` to return the shortfall
it encountered, which we use to populate the `shortfall` field.