Change `utxoAvailable` to type `UTxOSelection` in `SelectionParams`.
This commit allows callers of `CoinSelection.performSelection` to
pre-select a certain subset of inputs from the available UTxO set.

This pre-selection will form the basis of the input selection
created by the internal `Balance.performSelection` function.

While existing tests all still pass, further work will be required to
verify that `performSelection` behaves as expected if the set of
pre-selected UTxOs is non-empty.