Add `available{Balance,UTxO}` accessor functions to `UTxOSelection`.
The "available balance" is simply the sum of the selected and leftover
balances. It predicts what the selected balance would be if every single
UTxO were selected.

Similarly, the "available UTxO set" is the union of the selected and
leftover UTxO sets. It predicts what the selected UTxO set would be if
every single UTxO were selected.

The available balance and the available UTxO set both remain constant
over applications of `select` and `selectMany`.

This commit adds tests to verify that this is the case.