Split `utxoAvailable` into `utxoAvailableFor{Collateral,Inputs}`.
This allows each caller of `selectAssets` to make their own choice of:

- utxoAvailableForCollateral
- utxoAvailableForInputs

These sets are allowed to intersect: the ledger does not require that
they are disjoint.

However, some callers may prefer to pre-select a subset of inputs within
`utxoAvailableForInputs`, which is made possible with the `UTxOSelection`
`fromIndexFiltered` or `fromIndexPair` constructors.