Use type synonyms to simplify type signatures.
This commit arranges that we have:

- PerformSelectionOf inputId
- PerformSelection

- Selection{Error,Params,Result}Of inputId
- Selection{Error,Params,Result}

The variants with the `Of` suffix all have an `inputId` type parameter,
allowing the test suite to use whatever type of input identifier is most
convenient.

The variants without the `Of` suffix are all specialized to `TxIn`,
making it easier for external modules to use these types.

As a result of this commit, the top-level `CoinSelection` module will no
longer need to state `TxIn` in every type signature that uses a type
from the `Collateral` module.