Introduce function `reduceSelectionLimitBy`.
This function reduces a selection limit by a given amount.

We use this function to replace the incorrect subtraction in
`adjustComputeSelectionLimit`.

Background:

The `subtract` function (from `Prelude`) is equivalent to `flip (-)`,
which means that `subtract a b` is equivalent to `b - a`, rather than
the other way round.

Thanks to @KtorZ for discovering this.