Return all values that fail verification within `verifySelection`.
In cases where it's possible for multiple values to fail verification, our
current verification functions go to the effort of pattern matching on and
only returning the first such value that failed verification.

However, we can simplify the code by just returning all values that fail
verification.

If callers only wish to report the first value, then callers still have
the option to do this.