Cautiously use `RecordWildCards` to simplify verification functions.
Use of the `RecordWildCards` extension can lead to less readable code,
since it becomes harder for a reader to determine which fields have been
"absorbed" into a record value at the point of construction.

However, in the case of the post-condition verification functions
affected by this commit, it's relatively easy to tell which fields are
picked up, as the definition of each failure is always next to the
function that fails.

Therefore, we can safely use `RecordWildCards` to remove a bit of
verbosity.