Test variable parts of `InsufficientCollateral` API error.
It's important to test the variable parts of this error so that we we
can verify that the user is being presented with sensible advice.

It's also important to verify the numerical values as a sanity check
that they are sensible.

Note that our property tests all use simulated cost models, so it's
doubly important to have /something/ in the integration test suite to
act as a check for sensible values.