Add tests for long (with delegation) Shelley addresses
Strictly speaking, the Bech32 standard defines that the encoded sequence may not be longer than 90 characters — but Shelley addresses can be longer than that. We can cover this case with a test that looks more at functionality and considers addresses with delegation part.

Also changes suggested during review and additional tests.

Co-authored-by: Jonathan Knowles <jonathan.knowles@iohk.io>