Add preconditions to balanceTransaction
Using a not-yet-committed prop_balanceTransactionBalanced
it was discovered that balanceTransaction would produce
imbalanced transactions under several scenarios.

The guards/preconditions added here ensure the limitations
are clear to users who run into them. It also makes the
limitations clearly known and documented, allowing us to
come up with an appropriate strategy to fix.